Code cleanup and various small improvements
 - Moved Java project to use Maven as a build tool
 - The method Numbertext#load() now closes the streams properly
 - Changed variables to `final` where applicable
 - Using interfaces where applicable
 - General code autoformatting using Java conventions
