+++ java/.gitignore
@@ -1,0 +1,5 @@
+.classpath
+.project
+.settings/
+*.class
+/target/
