+++ java/src/main/java/org/numbertext/MenuState.java
@@ -1,0 +1,5 @@
+package org.numbertext;
+
+public enum MenuState {
+	PARAM, LANGUAGE, PREFIX;
+}
